+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Mobs now flee from the warden instead of attaking it
+
 ## 1.15.4
 * Fixed world freezing when a Wither Skeleton loaded